+++ pydidas/gui/frames/builders/data_browsing_frame_builder.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -19,17 +21,17 @@ the DataBrowsingFrame with widgets.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["DataBrowsingFrameBuilder"]
 
-from qtpy import QtWidgets, QtCore
 import qtawesome as qta
+from qtpy import QtCore, QtWidgets
 
 from ....core import constants
-from ....widgets import BaseFrame
+from ....widgets.framework import BaseFrame
 from ....widgets.selection import DirectoryExplorer, Hdf5DatasetSelector
 from ....widgets.silx_plot import PydidasImageView