+++ pydidas/gui/frames/builders/define_diffraction_exp_frame_builder.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -19,17 +21,16 @@ populate the DefineDiffractionExpFrame with widgets.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["DefineDiffractionExpFrameBuilder"]
 
 from ....core import constants
-from ....core.constants import (
-    FIX_EXP_POLICY,
-)
-from ....widgets import ScrollArea, BaseFrame
+from ....core.constants import FIX_EXP_POLICY
+from ....widgets import ScrollArea
+from ....widgets.framework import BaseFrame
 from ....widgets.parameter_config import ParameterEditCanvas