+++ pydidas/gui/windows/pydidas_window.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -19,24 +21,23 @@ to is a QMainWindow widget and is used to extend BaseFrames to stand-alone Qt wi
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["PydidasWindow"]
 
 import os
 
-from qtpy import QtCore, QtWidgets, QtGui
+from qtpy import QtCore, QtGui, QtWidgets
 
 from ...core.utils import (
-    get_pydidas_icon_w_bg,
-    doc_qurl_for_window_manual,
-    doc_filename_for_window_manual,
     DOC_HOME_QURL,
+    doc_filename_for_window_manual,
+    doc_qurl_for_window_manual,
+    get_pydidas_icon_w_bg,
 )
-
-from ...widgets import BaseFrame
+from ...widgets.framework import BaseFrame
 
 
 class PydidasWindowMixIn: