+++ pydidas/widgets/silx_plot/silx_plot_stack_factory.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -14,12 +16,13 @@
 # along with Pydidas. If not, see <http://www.gnu.org/licenses/>.
 
 """
-Module with a function to create a QStackedWidget with 1d and 2d plots.
+Module with a function to create a QStackedWidget with 1d and 2d plots and add it to
+a Pydidas Frame.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["create_silx_plot_stack"]
@@ -40,7 +43,7 @@ def create_silx_plot_stack(frame, gridPos=None, use_data_info_action=False):
 
     Parameters
     ----------
-    frame : pydidas.core.BaseFrame
+    frame : pydidas.widgets.framework.BaseFrame
         The input frame.
     gridPos : Union[tuple, None], optional
         The gridPos for the new widget. The default is None.
@@ -50,7 +53,7 @@ def create_silx_plot_stack(frame, gridPos=None, use_data_info_action=False):
 
     Returns
     -------
-    frame : pydidas.core.BaseFrame
+    frame : pydidas.widgets.framework.BaseFrame
         The updated frame.
     """
     frame._widgets["plot1d"] = PydidasPlot1D()